pr-feedback/requirements.txt
PyGithub
google-cloud-aiplatform
google-generativeai